+++ src/components/icon/Icon.snap.ts
@@ -1,0 +1,31 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`matches the snapshot Case 0 With icon: ({"icon": "tag"}) => HTML 1`] = `
+<span
+  class="wvui-icon oo-ui-icon-tag"
+>
+  
+	
+
+</span>
+`;
+
+exports[`matches the snapshot Case 1 With icon, inverted: ({"icon": "tag", "invert": true}) => HTML 1`] = `
+<span
+  class="wvui-icon oo-ui-icon-tag oo-ui-image-invert"
+>
+  
+	
+
+</span>
+`;
+
+exports[`matches the snapshot Case 2 With icon and label: ({"icon": "tag", "label": "Tag this image"}) => HTML 1`] = `
+<span
+  class="wvui-icon oo-ui-icon-tag"
+>
+  
+	Tag this image
+
+</span>
+`;